Remove row_mysql_freeze_data_dictionary()
The SQL layer will acquire MDL on (grand)child tables,
therefore it should be safe to dereference the pointers
in dict_table_t::referenced_set.

Because the SQL layer is holding MDL on the current table,
it is also safe to access (but not necessarily dereference the pointers in)
dict_table_t::foreign_set.

Remove unnecessary dict_sys.latch acquisition.